+++ tests/test_fastquant.py
@@ -1,5 +1,6 @@
 import pandas as pd
 from fastquant import (
+    get_disclosures_df,
     get_pse_data,
     get_yahoo_data,
     get_stock_data,
@@ -11,6 +12,14 @@ PHISIX_SYMBOL = "JFC"
 YAHOO_SYMBOL = "GOOGL"
 DATE_START = "2018-01-01"
 DATE_END = "2019-01-01"
+
+
+def test_get_disclosures_df():
+    df_dict = get_disclosures_df(PHISIX_SYMBOL, DATE_START, DATE_END)
+    dfd = df_dict["D"]
+    dfe = df_dict["E"]
+    assert isinstance(dfd, pd.DataFrame)
+    assert isinstance(dfe, pd.DataFrame)
 
 
 def test_get_pse_data():
